update refactor code to decouple dom from logic
